requirements.txt
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037869401/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1626470256873/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831756778/work
click==8.0.1
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877477759/work
ipykernel @ file:///C:/ci/ipykernel_1607454116140/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617120911285/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.0.1
jedi @ file:///C:/ci/jedi_1606932640618/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
joblib==1.0.1
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213515523/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver==1.3.1
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
matplotlib==3.4.2
mistune @ file:///C:/ci/mistune_1607359457024/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479160025/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1621625801656/work
numpy==1.21.1
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandas==1.3.1
pandocfilters @ file:///C:/ci/pandocfilters_1605114832805/work
parso==0.7.0
patsy==0.5.1
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.3.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1623189609245/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1625052426389/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.1
pywin32==228
pywinpty @ file:///C:/ci/pywinpty_1607419945780/work
pyzmq==20.0.0
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
statsmodels==0.12.2
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==2.2.0
tornado @ file:///C:/ci/tornado_1606924294691/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.1
wincertstore==0.2
xgboost==1.4.2
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
